Fixed error where ExprFunc used an uninitialized factorIntMat.
This totally invalidates previous results from the "new_storage" branch.
I alwasy test new branches against old branches though, so this may hint
that the output was matching older versions. Maybe a similar problem
existed before. >.<

The reason this is a problem is that code does not check to see which
factors intereact before checking factorIntMat, a non-interaction should
be 1.0 and the code just expects those elements of factorIntMat to
contain 1.0.

Initialize to 1.0, then populate.