Fixed calculation of distances and overlaps.
This will now produce the same results on the reverse complement
sequence. Models will have to have their distances altered to be the
same model.

I used some macros to calculate distances quickly. It is imperative that
sites be presented to the interaction functions and to the macros in the
right order. This _may_ be faster.

Turn on (and add) some debugging asserts to check any new DP, then turn
them off again once your DP is known to be correct.